for some reason, the parser keeps reading the angle column from rotate 180 as a string. I'm sick of dealing with Excel's formatting nonsense, so I'll move onto something else before trying to fix this

+++ src/main/java/autocadDrawingChecker/data/core/RecordExtractor.java
@@ -38,7 +38,7 @@ public class RecordExtractor {
                 ret = c.getBooleanCellValue();
                 break;
             case NUMERIC:
-                ret = c.getNumericCellValue();
+                ret = (Double)c.getNumericCellValue();
                 break;
             case STRING:
                 ret = c.getStringCellValue();
